Convert ArticleImageSet to compound component
